+++ modules/hardware/default.nix
@@ -1,5 +1,4 @@
 {
-  config,
   inputs,
   lib,
   util,
@@ -8,31 +7,12 @@
 let
   inherit (util.map) modules;
   inherit (lib) mkForce mkOption types;
-  cfg = config.hardware;
 in
 {
   ## HARDWARE Configuration ##
   imports = modules.list ./.;
 
   options.hardware = with types; {
-    cpu = {
-      cores = mkOption {
-        description = "Number of CPU Cores";
-        type = int;
-        default = 4;
-      };
-
-      mode = mkOption {
-        description = "CPU Frequency Governor Mode";
-        type = enum [
-          "ondemand"
-          "performance"
-          "powersave"
-        ];
-        default = "performance";
-      };
-    };
-
     modules = mkOption {
       description = "List of Modules imported from 'inputs.hardware'";
       type = listOf (enum (attrNames inputs.hardware.nixosModules));
@@ -47,9 +27,6 @@ in
   };
 
   config = {
-    powerManagement.cpuFreqGovernor = cfg.cpu.mode;
-    nix.settings.max-jobs = cfg.cpu.cores;
-
     specialisation.powersave.configuration = {
       system.nixos.label = "special.powersave";
       gui.fancy = mkForce false;